Extract the duplicated PNG container writer
QRCodePngRenderer and BarcodePngRenderer held byte-identical copies of the PNG
signature, the four chunk type arrays, CompressImageData, WritePng, WriteChunk,
ComputeCrc32, UpdateCrc32 and InitializeCrc32Table - about 120 lines each, including
two separate 256-entry CRC tables built and held for the process lifetime.

Moves the container writing into an internal PngWriter. The scanline generation stays
in each renderer because the two differ: QR codes carry a quiet zone on all four
sides, barcodes only on the left and right.

No behaviour change - every PNG snapshot is byte identical.
